Add notification and refactor
