+++ app/javascript/components/SlotPage.js
@@ -1,11 +1,11 @@
 import React, { useState } from "react";
+import { getList } from "../services/SlotApi";
+import SlotBooking from "../components/SlotBooking";
 
 const SlotPage = () => {
   const [date, setDate] = useState("");
   const [time, setTime] = useState("");
   const [list, setList] = useState([]);
-
-  // ------------- GET
 
   const getDurationSlots = () => {
     let mounted = true;
@@ -17,62 +17,10 @@ const SlotPage = () => {
     return () => (mounted = false);
   };
 
-  const getList = (date, time) => {
-    return fetch(
-      `http://localhost:3000/slots?date=${date}T00:00:00+00:00&minutes=${time}`
-    ).then((data) => data.json());
-  };
-
-  const formatDate = (date) => {
-    let time_selected = date.slice(11, 16);
-    return time_selected;
-  };
-
   const handleSubmit = (e) => {
     e.preventDefault();
     getDurationSlots(date, time);
   };
-
-  // -------------- PUT
-  const [duration, setDuration] = useState("");
-
-  const bookDurationSlot = (duration) => {
-    let mounted = true;
-    bookSlot(duration).then(() => {
-      if (mounted) {
-        console.log(duration, "duration");
-      }
-    });
-    return () => (mounted = false);
-  };
-
-  const bookSlot = (duration) => {
-    return fetch(
-      `http://localhost:3000/slots?start_date=${duration.start_time}&end_date=${duration.end_time}`,
-      {
-        method: "PUT",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      }
-    )
-      .then((data) => data.json())
-      .catch((error) => {
-        console.log("Error:", error);
-      });
-  };
-
-  const handleSlotSelection = (e) => {
-    e.preventDefault();
-    bookDurationSlot(duration);
-    showSuccessNotification();
-  };
-
-  const showSuccessNotification = () => {
-    
-  }
-
-  // ------------
 
   return (
     <div>
@@ -88,36 +36,18 @@ const SlotPage = () => {
           <input type="date" onChange={(e) => setDate(e.target.value)} />
           <input
             type="input"
-            pattern="[+-]?\d+(?:[.,]\d+)?"
             placeholder="Enter time in minutes"
             onChange={(e) => setTime(e.target.value)}
           />
         </label>
         <input type="submit" value="Submit" />
       </form>
-      {list.length !== 0 && (
-        <form
-          onSubmit={(e) => {
-            handleSlotSelection(e);
-          }}
-        >
-          <div>
-            <h1>Available Slot List </h1>
-            <h3>Please select a slot 👇 </h3>
-            <ul>
-              {list.map((item, index) => (
-                <button
-                  key={index}
-                  index={index}
-                  onClick={() => setDuration(item)}
-                >
-                  {formatDate(item.start_time)} - {formatDate(item.end_time)}
-                </button>
-              ))}
-            </ul>
-          </div>
-        </form>
-      )}
+      <SlotBooking
+        list={list}
+        getDurationSlots={getDurationSlots}
+        date={date}
+        time={time}
+      />
     </div>
   );
 };
